Center Content and Semantic Elements
The content was centered by adding min-height.

max-width was used with rem units instead of width for better responsiveness.

Semantic elements were used.

The content was put into the main tag and the use of extra class names was avoided.

Padding value of body was changed for better visual on small screens.

New screenshot added.